Create traits directory, relocate auction-manager, create bc-country trait
