+++ pallets/auction/Cargo.toml
@@ -37,7 +37,7 @@ primitives = { package = "bit-country-primitives", path = "../primitives", defau
 orml-nft = { default-features = false, version = '0.4.0' }
 pallet-nft = { default-features = false, package = 'pallet-nft', path = '../nft' }
 pallet-continuum = { default-features = false, package = 'pallet-continuum', path = '../continuum' }
-auction-manager = { default-features = false, package = 'auction-manager', path = '../auction-manager' }
+auction-manager = { default-features = false, package = 'auction-manager', path = '../../traits/auction-manager' }
 
 
 [features]
