+++ pallets/bitcountry/Cargo.toml
@@ -30,6 +30,9 @@ sp-session = { version = "3.0.0", default-features = false }
 sp-version = { version = "3.0.0", default-features = false }
 primitives = { package = "bit-country-primitives", path = "../primitives", default-features = false }
 
+# local packages
+bc-country = { path = "../../traits/bc-country", default-features = false }
+
 # Orml packages
 orml-traits = { default-features = false, version = '0.4.0' }
 orml-currencies = { default-features = false, version = '0.4.0' }
@@ -40,7 +43,6 @@ default-features = false
 package = 'pallet-nft'
 path = '../nft'
 version = '2.0.0-rc6'
-
 
 [features]
 default = ['std']
