+++ traits/auction-manager/Cargo.toml
@@ -1,0 +1,32 @@
+[package]
+authors = ['Bit Country <https://github.com/bit-country>']
+description = 'Bit Country pallet for auction management.'
+edition = '2018'
+homepage = 'https://bit.country'
+license = 'Unlicense'
+name = 'auction-manager'
+repository = 'https://github.com/bit-country'
+version = '2.0.0-rc6'
+
+[dependencies]
+serde = { version = "1.0.119", optional = true, features = ["derive"] }
+codec = { package = "parity-scale-codec", version = "2.0.0", default-features = false }
+sp-runtime = { version = "3.0.0", default-features = false }
+sp-std = { version = "3.0.0", default-features = false }
+sp-io = { version = "3.0.0", default-features = false }
+frame-support = { version = "3.0.0", default-features = false }
+# Used for the node's RPCs
+primitives = { package = "bit-country-primitives", path = "../../pallets/primitives", default-features = false }
+
+funty = { version = "=1.1.0", default-features = false }
+
+[features]
+default = ['std']
+std = [
+    'serde',
+    'codec/std',
+    'sp-runtime/std',
+    'sp-io/std',
+    'sp-std/std',
+    'frame-support/std',
+]
