+++ traits/auction-manager/src/lib.rs
