+++ traits/bc-country/Cargo.toml
@@ -1,10 +1,10 @@
 [package]
 authors = ['Bit Country <https://github.com/bit-country>']
-description = 'Bit Country pallet for auction management.'
+description = 'Bit Country pallet for block logic.'
 edition = '2018'
 homepage = 'https://bit.country'
 license = 'Unlicense'
-name = 'auction-manager'
+name = 'bc-country'
 repository = 'https://github.com/bit-country'
 version = '2.0.0-rc6'
 
@@ -15,10 +15,7 @@ sp-runtime = { version = "3.0.0", default-features = false }
 sp-std = { version = "3.0.0", default-features = false }
 sp-io = { version = "3.0.0", default-features = false }
 frame-support = { version = "3.0.0", default-features = false }
-# Used for the node's RPCs
-primitives = { package = "bit-country-primitives", path = "../primitives", default-features = false }
-
-funty = { version = "=1.1.0", default-features = false }
+primitives = { package = "bit-country-primitives", path = "../../pallets/primitives", default-features = false }
 
 [features]
 default = ['std']
